bimtester: steps, add new test module for spatial structure
